PCI precision, radial symmetry, and more precision in identifier hashes

+++ trimesh/triangles.py
@@ -161,7 +161,8 @@ def mass_properties(triangles,
                                            (triangles[:, 1, triangle_i] * g1[:, i]) +
                                            (triangles[:, 2, triangle_i] * g2[:, i]))
 
-    coefficents = 1.0 / np.array([6, 24, 24, 24, 60, 60, 60, 120, 120, 120])
+    coefficents = 1.0 / np.array([6, 24, 24, 24, 60, 60, 60, 120, 120, 120],
+                                 dtype=np.float64)
     integrated = integral.sum(axis=1) * coefficents
 
     volume = integrated[0]
